Simplify logic
When you find yourself writing:

```ruby
if condition
  true
else
  false
end
```

you can erase the whole if statement. Remember, if runs the first branch
if the condition is true. In this case, that makes it return true, which
is what the conditional is already returning.

`completed == false` is only necessary if you want to ensure complete
isn't nil. Otherwise, `!completed` is the better way to do this.